feat(auth): optional support of restricted groups
LDAPAuthentifier now accepts an optional list of restricted groups. Upon
user login, LDAPAuthentifier raises an LDAPAuthenticationError if
restricted groups are set and the user is not member of any of these
groups. The users() method also discard users not member of restricted
groups when users are retrieved with groups memberships and restricted
groups are set.