feat(settings): support parameters deprecation
Add a way to mark parameters as deprecated with new replacement
parameter.

When a deprecated parameter is defined in site configuration, RFL checks
the replacement parameter is not defined as well. In this case, a
warning log is emitted to mention the deprecated parameter is ingored.
When a deprecated parameter is actually loaded FutureWarning is raised
to warn end user.

fix #48